fix(analytics): move mono/HELP out of the client module
The admin page (a server component) called mono() and read HELP from
"use client" panels.tsx — client-module exports become client
references on the server, and invoking one crashes at request time
("Attempted to call mono() from the server"). Neither tsc, jest, nor
next build catches this on a dynamic page.

mono and HELP now live in style.ts (no directive); panels re-exports
them so client importers are unchanged.
